requirements.txt
asgiref==3.8.1
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
inflection==0.5.1
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
gunicorn==23.0.0
whitenoise==6.9.0
psycopg2-binary==2.9.10
dj-database-url==3.0.0
django-jazzmin==3.0.1
cryptography==45.0.5
requests==2.32.4
firebase-admin==6.9.0
couchbase==4.4.0
django-redis==6.0.0
google-analytics-data==0.18.19
libcouchbase-dev